+++ packages/markdown-core/src/index.ts
@@ -3,3 +3,4 @@ export * from './lib/parse';
 export * from './lib/serialize';
 export * from './lib/transform';
 export * from './lib/id';
+export * from './lib/color-span';